+++ procfs/Cargo.toml
@@ -22,7 +22,6 @@ rustix = { version = "0.37.0", features = ["fs", "process", "param", "thread"] }
 bitflags = "1.2"
 lazy_static = "1.0.2"
 chrono = {version = "0.4.20", optional = true, features = ["clock"], default-features = false }
-byteorder = {version="1.2.3", features=["i128"]}
 hex = "0.4"
 flate2 = { version = "1.0.3", optional = true }
 backtrace = { version = "0.3", optional = true }
