+++ procfs/examples/pressure.rs
@@ -1,7 +1,8 @@
-/// A basic example of /proc/pressure/ usage.
+use procfs::prelude::*;
 
+/// A basic example of /proc/pressure/ usage.
 fn main() {
-    println!("memory pressure: {:#?}", procfs::MemoryPressure::new());
-    println!("cpu pressure: {:#?}", procfs::CpuPressure::new());
-    println!("io pressure: {:#?}", procfs::IoPressure::new());
+    println!("memory pressure: {:#?}", procfs::MemoryPressure::current());
+    println!("cpu pressure: {:#?}", procfs::CpuPressure::current());
+    println!("io pressure: {:#?}", procfs::IoPressure::current());
 }
